feat: separate shape related functions
