+++ perception/autoware_multi_object_tracker/CMakeLists.txt
@@ -31,6 +31,7 @@ set(${PROJECT_NAME}_lib
   lib/association/association.cpp
   lib/association/mu_successive_shortest_path/mu_ssp.cpp
   lib/object_model/dynamic_object.cpp
+  lib/object_model/shapes.cpp
   lib/tracker/motion_model/motion_model_base.cpp
   lib/tracker/motion_model/bicycle_motion_model.cpp
   # cspell: ignore ctrv
