+++ perception/autoware_multi_object_tracker/lib/tracker/motion_model/cv_motion_model.cpp
@@ -21,12 +21,12 @@
 
 #include "autoware/multi_object_tracker/tracker/motion_model/motion_model_base.hpp"
 #include "autoware/multi_object_tracker/utils/utils.hpp"
-#include "autoware/universe_utils/math/normalization.hpp"
-#include "autoware/universe_utils/math/unit_conversion.hpp"
-#include "autoware/universe_utils/ros/msg_covariance.hpp"
 
 #include <Eigen/Core>
 #include <Eigen/Geometry>
+#include <autoware/universe_utils/math/normalization.hpp>
+#include <autoware/universe_utils/math/unit_conversion.hpp>
+#include <autoware/universe_utils/ros/msg_covariance.hpp>
 
 #include <tf2/utils.h>
