+++ perception/autoware_multi_object_tracker/src/processor/processor.cpp
@@ -16,6 +16,7 @@
 
 #include "autoware/multi_object_tracker/object_model/dynamic_object.hpp"
 #include "autoware/multi_object_tracker/object_model/object_model.hpp"
+#include "autoware/multi_object_tracker/object_model/shapes.hpp"
 #include "autoware/multi_object_tracker/tracker/tracker.hpp"
 
 #include <autoware/object_recognition_utils/object_recognition_utils.hpp>
